use new type Movie in backend functions
