+++ libs/utils/src/lib/helpers.ts
@@ -27,27 +27,6 @@ export function keyToAddressPart(key: Key, length: number): AddressParts {
   return {start: address.slice(0, length), middle: address.slice(length, address.length - length), end: address.slice(-length)};
 }
 
-export interface DateRangeRaw<D> {
-  from: D;
-  to: D
-}
-
-/** Need it for calendar components */
-export interface DateRange extends DateRangeRaw<Date> {
-}
-
-export function createDateRange(params: Partial<DateRange> = {}): DateRange {
-  return {
-    from: null,
-    to: null,
-    ...params
-  }
-}
-
- /** check if a date is in a range */
- export function isBetween(date: Date, startRange: Date, endRange: Date){
-  return date.getTime() >= startRange.getTime() && date.getTime() <= endRange.getTime();
-}
 /** Get first part of an ens domain : `alice.blockframes.eth` -> `alice` */
 export function getNameFromENS(ensDomain: string) {
   return ensDomain.split('.')[0];
