+++ tsconfig.json
@@ -41,6 +41,7 @@
       "@blockframes/auth/types": ["libs/auth/src/lib/+state/auth.firestore.ts"],
       "@blockframes/invitation/types": ["libs/notification/src/lib/invitation/+state/invitation.firestore.ts"],
       "@blockframes/material/material/types": ["libs/material/src/lib/material/+state/material.firestore.ts"],
+      "@blockframes/movie/types": ["libs/movie/src/lib/movie/+state/movie.firestore.ts"],
 
       "@blockframes/account": ["libs/account/src/index.ts"],
